Add cli command
This is a very basic cli, but could be enough for most purposes.

When installed globally, it can then be used as `dotenv <normal command>` and then the normal command will have the necessary environment variables.

Please let me know if anything needs changing. Or if this should be a seperate module.