Return boolean values instead of "true"/"false" strings.
When boolean values are set in an .env file (e.g. IS_PRODUCTION=false), it's expected that these will be evalulated as booleans, instead of the string "false" (which is truthy).  This causes particular problems when using webpack and webpack.DefinePlugin, since providing `JSON.stringify(process.env.IS_PRODUCTION` will evaluate as a string, which is unexpected.

Instead, change `"true"` to `true` and `"false"` to `false` when evaluating the .env file.