docs: update readme with Fastify reference application
This PR updates the README.md file with a reference to a fully working Fastify application that uses dotenv.

This should only merge once the PR (https://github.com/dotenv-org/examples/pull/6) on the `dotenv-org/examples` merges too. It also has more insights and information as to the benefits of the referenced Fastify app with dotenv.